rename test_deck_ballots.* files to test_decks.*
These files now cover more than just ballots wrt. test decks
